+++ apps/simulation-study-aggregated-observations/Main.hs
@@ -27,7 +27,7 @@ import BDSCOD.Types
   , scheduledTimes
   , unpackParameters
   )
-import BDSCOD.Utility (eventsAsObservations,invLogit)
+import BDSCOD.Utility (eventsAsObservations, invLogit)
 import Control.Monad (liftM2, when)
 import Control.Monad.Except (ExceptT, runExceptT, throwError)
 import Control.Monad.Reader (ReaderT, asks, liftIO, runReaderT)
@@ -37,6 +37,7 @@ import qualified Data.ByteString.Lazy as L
 import qualified Data.Csv as Csv
 import Data.List (intersperse)
 import Data.Maybe (fromJust, isJust)
+import qualified Data.Vector.Unboxed as Unboxed
 import qualified Epidemic.BDSCOD as SimBDSCOD
 import Epidemic.Types.Events
   ( EpidemicEvent(..)
@@ -50,9 +51,9 @@ import qualified Epidemic.Utility as SimUtil
 import GHC.Generics
 import Numeric.GSL.Minimization (MinimizeMethod(NMSimplex2), minimizeV)
 import Numeric.LinearAlgebra.Data (Vector(..), fromList, linspace, toList)
-
 -- import Numeric.LinearAlgebra.HMatrix
 import System.Environment (getArgs)
+import System.Random.MWC (initialize)
 
 -- | These objects define the specifics of the evaluation of LLHD profiles. If a
 -- point estimate is given, then that is the central point of the profiles,
@@ -144,11 +145,14 @@ bdscodConfiguration = do
 
 -- | Simulate the actual epidemic making sure that the results are acceptable
 -- before returning the results.
-simulateEpidemic bdscodConfig = do
+--
+-- When the simulation fails this attempts to simulate the data set again useing
+-- a different seed to avoid getting stuck in a loop.
+simulateEpidemic seedInt bdscodConfig = do
   ifVerbosePutStrLn "Running simulateEpidemic..."
+  genIO <- liftIO $ initialize (Unboxed.fromList [seedInt])
   simEvents <-
-    liftIO $
-    SimUtil.simulationWithSystemRandom False bdscodConfig SimBDSCOD.allEvents
+    liftIO $ SimUtil.simulation' bdscodConfig SimBDSCOD.allEvents genIO
   (sizeLowerBound, sizeUpperBound) <- asks simulationSizeBounds
   if length simEvents > sizeLowerBound && length simEvents < sizeUpperBound
     then do
@@ -159,7 +163,7 @@ simulateEpidemic bdscodConfig = do
       return simEvents
     else do
       ifVerbosePutStrLn "\trepeating the simulation..."
-      simulateEpidemic bdscodConfig
+      simulateEpidemic (succ seedInt) bdscodConfig
 
 -- | Take a simulated epidemic and generate the observations, first with full
 -- resolution of the event times and the true epidemic parameters, second with
@@ -297,9 +301,12 @@ evaluateLLHD infConfig obs = do
 -- | Using regular (i.e., disaggregated) observations, estimate the parameters
 -- and evaluate the log-likelihood profiles and write the result to file. This
 -- will also evaluate the density of the prevalence at the present and write
--- that to file. __NOTE__ This uses the actual simulation parameters as a way to
--- get the scheduled observation times, they are not used in the inference, that
--- starts at a fixed initial condition.
+-- that to file.
+--
+-- __NOTE__ This uses the actual simulation parameters as a way to get the
+-- scheduled observation times, they are not used in the inference, that starts
+-- at a fixed initial condition.
+--
 estimateLLHD :: InferenceConfiguration -> [Observation] -> Simulation ()
 estimateLLHD infConfig obs = do
   liftIO (putStrLn "Running estimateLLHD...")
@@ -331,7 +338,8 @@ estimateLLHDAggregated infConfig (AggregatedObservations (AggTimes aggTimes) obs
 simulationStudy :: Simulation ()
 simulationStudy = do
   bdscodConfig <- bdscodConfiguration
-  epiSim <- simulateEpidemic bdscodConfig
+  let seedInt = 42
+  epiSim <- simulateEpidemic seedInt bdscodConfig
   infConfigs <- asks inferenceConfigurations
   (regObs, regObs', aggObs) <- observeEpidemicThrice epiSim infConfigs
   uncurry evaluateLLHD regObs